+++ CHANGELOG.md
@@ -8,7 +8,11 @@ Please see [CONTRIBUTING.md](https://github.com/cucumber/cucumber/blob/master/CO
 
 ### Added
 
-* N/A
+* Added rubocop (with todo file), and removed backports gems
+  ([#186](https://github.com/cucumber/cucumber-ruby-core/pull/186),
+  [#182](https://github.com/cucumber/cucumber-ruby-core/issues/182)
+  [tas50](https://github.com/tas50),
+  [luke-hill](https://github.com/luke-hill))
 
 ### Fixed
 
@@ -21,7 +25,6 @@ Please see [CONTRIBUTING.md](https://github.com/cucumber/cucumber/blob/master/CO
 ### Improved
 
 * N/A
-
 
 ## [5.0.2](https://github.com/cucumber/cucumber-ruby-core/compare/v5.0.1...v5.0.2)